backend/apps/atlas/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from . import views

router = DefaultRouter()
router.register(r'facilities', views.FacilityViewSet, basename='facility')
router.register(r'floors', views.FloorViewSet, basename='floor')
router.register(r'spots', views.ParkingSpotViewSet, basename='spot')
router.register(r'devices', views.DeviceViewSet, basename='device')

urlpatterns = [
    path('', include(router.urls)),
]
