backend/pyproject.toml
[project]
name = "parkhero"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "qrcode>=8.2",
]
